Split prompt planning and rendering
