+++ Brainarr.Plugin/Services/Prompting/IPromptRenderer.cs
@@ -1,0 +1,8 @@
+using System.Threading;
+
+namespace NzbDrone.Core.ImportLists.Brainarr.Services.Prompting;
+
+public interface IPromptRenderer
+{
+    string Render(PromptPlan plan, ModelPromptTemplate template, CancellationToken cancellationToken);
+}
