+++ Brainarr.Plugin/Services/Prompting/ModelPromptTemplate.cs
@@ -1,0 +1,6 @@
+namespace NzbDrone.Core.ImportLists.Brainarr.Services.Prompting;
+
+public sealed record ModelPromptTemplate(string Name)
+{
+    public static ModelPromptTemplate Default { get; } = new("default");
+}
